healog/src/main/java/Genesis/healog/service/MemberService.java
package Genesis.healog.service;

import Genesis.healog.domain.Member;
import Genesis.healog.repogitory.MemberRepository;
import Genesis.healog.repogitory.MemoryMemberRepository;

import java.util.List;
import java.util.Optional;

public class MemberService {
    private final MemberRepository memberRepository = new MemoryMemberRepository();

    // 회원 가입

    public Long join(Member member) {
        validateDuplicateMember(member); // 중복 회원 X
        memberRepository.save(member);
        return member.getId();
    }

    private void validateDuplicateMember(Member member) {
        memberRepository.findByName(member.getName())
                .ifPresent(m-> {
                    throw new IllegalStateException("이미 존재하는 회원입니다.");
                });
    }

    public List<Member> findAll() {
        return memberRepository.findAll();
    }

    public Optional<Member> findOne(Long id) {
        return memberRepository.findById(id);
    }
}
